chat-app/requirements.txt
grpcio==1.53.0
mysql-connector-python==8.0.32
confluent-kafka==2.1.0
